Move sort and filter state into a single uiReducer
