+++ frontend/src/store.js
@@ -1,14 +1,12 @@
 import { configureStore } from '@reduxjs/toolkit'
 import userReducer from './reducers/userReducer'
 import feedbacksReducer from './reducers/feedbacksReducer'
-import filterReducer from './reducers/filterReducer'
-import sortReducer from './reducers/sortReducer'
+import uiReducer from './reducers/uiReducer'
 
 export const store = configureStore({
   reducer: {
     currentUser: userReducer,
     productRequests: feedbacksReducer,
-    filter: filterReducer,
-    sort: sortReducer,
+    ui: uiReducer,
   },
 })
